+++ app/src/features/apiClient/slices/buffer/thunks.ts
@@ -8,7 +8,6 @@ import { createRecord, updateRecord } from "../apiRecords/thunks";
 import { ApiClientEntityType } from "../entities/types";
 import { BUFFER_SLICE_NAME } from "../common/constants";
 import type { BufferEntry } from "./types";
-import { emitBufferUpdated } from "componentsV2/Tabs/slice/previewTabManager";
 
 type Repository = ApiClientRecordsInterface<Record<string, unknown>>;
 
@@ -85,11 +84,6 @@ export const saveBuffer = createAsyncThunk<
       referenceId: result.id,
     })
   );
-
-  emitBufferUpdated({
-    entityId: result.id,
-    entityType: entry.entityType,
-  });
 
   return result;
 });
